+++ explorer_launch_other.go
@@ -1,0 +1,11 @@
+// SPDX-FileType: SOURCE
+// SPDX-License-Identifier: BSD-3-Clause
+// Project: https://github.com/woozymasta/flags
+
+//go:build !windows
+
+package flags
+
+func detectLaunchedFromExplorer() bool {
+	return false
+}